@@ -5,11 +5,11 @@ description: Flux library for uni-directional dataflow inspired by reflux and Fa
 homepage: https://github.com/Workiva/w_flux
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   meta: ^1.6.0
-  react: ">=5.7.0 <7.0.0"
+  react: ^7.0.0
   w_common: ^3.0.0
 
 dev_dependencies:
